+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.7.9 (2022-10-15)
+- Implement row expansion [controlled mode](https://icflorescu.github.io/mantine-datatable/examples/row-expansion-examples)
+- Implement margin props on the DataTable component (m, my, mx, mt, mb, ml, mr)
+
 ## 1.7.8 (2022-10-14)
 - Add [cell click handlers](https://icflorescu.github.io/mantine-datatable/examples/handling-cell-clicks)
 - Add `recordIndex` as a second parameter on various hadlers and callbacks (such as `onRowClick`, column `render`, etc.)